Added min_size arg to change only huge files if needed

+++ img-to-webp.py
@@ -4,11 +4,14 @@ import argparse
 from PIL import Image
 
 
-def converter(image_name):
-    img = Image.open(image_name).convert('RGB')
-    img.save(os.path.splitext(image_name)[
-             0] + '.webp', 'webp', optimize=True)
-    os.remove(image_name)
+def converter(image_name, min_size):
+    if min_size is None or os.path.getsize(image_name) > min_size * 1024:
+        print(f"Working on {image_name}...")
+        img = Image.open(image_name).convert('RGB')
+        img.save(os.path.splitext(image_name)[
+            0] + '.webp', 'webp', optimize=True, quality=80)
+        if args["remove_originals"]:
+            os.remove(image_name)
 
 
 if __name__ == '__main__':
@@ -18,12 +21,16 @@ if __name__ == '__main__':
                         type=str, default=os.getcwd() + '/')
     parser.add_argument('--image_name', '-n',
                         type=str, default=None)  # for single images
+    parser.add_argument('--remove_originals', '-r',
+                        type=bool, default=False)
+    parser.add_argument('--min_size', '-m',
+                        type=int, default=None)  # in kBs
 
     args = vars(parser.parse_args())
 
     if args["image_name"]:
         if os.path.isfile(args["image_name"]):
-            converter(args["image_name"])
+            converter(args["image_name"], args["min_size"])
         else:
             print("Image path does not exist!")
     else:
@@ -31,7 +38,7 @@ if __name__ == '__main__':
             args["image_directory"] += '/'
 
         files = [file for file in os.listdir(args["image_directory"]) if re.match(
-            ".*\.(?:jpg|jpeg|png)", file)]
+            ".*\.(?:jpg|jpeg|png|webp)", file)]
 
         for file in files:
-            converter(args["image_directory"] + file)
+            converter(args["image_directory"] + file, args["min_size"])
